@@ -11,10 +11,11 @@
 
 from typing import TYPE_CHECKING
 
+from msrest import Deserializer, Serializer
+
 from azure.mgmt.core import ARMPipelineClient
 from azure.profiles import KnownProfiles, ProfileDefinition
 from azure.profiles.multiapiclient import MultiApiClientMixin
-from msrest import Deserializer, Serializer
 
 from ._configuration import CustomLocationsConfiguration
 
@@ -23,7 +24,6 @@ if TYPE_CHECKING:
     from typing import Any, Optional
 
     from azure.core.credentials import TokenCredential
-    from azure.core.pipeline.transport import HttpRequest, HttpResponse
 
 class _SDKClient(object):
     def __init__(self, *args, **kwargs):
@@ -70,12 +70,10 @@ class CustomLocations(MultiApiClientMixin, _SDKClient):
         credential,  # type: "TokenCredential"
         subscription_id,  # type: str
         api_version=None, # type: Optional[str]
-        base_url=None,  # type: Optional[str]
+        base_url="https://management.azure.com",  # type: str
         profile=KnownProfiles.default, # type: KnownProfiles
         **kwargs  # type: Any
     ):
-        if not base_url:
-            base_url = 'https://management.azure.com'
         self._config = CustomLocationsConfiguration(credential, subscription_id, **kwargs)
         self._client = ARMPipelineClient(base_url=base_url, config=self._config, **kwargs)
         super(CustomLocations, self).__init__(
@@ -93,12 +91,16 @@ class CustomLocations(MultiApiClientMixin, _SDKClient):
 
            * 2021-03-15-preview: :mod:`v2021_03_15_preview.models<azure.mgmt.extendedlocation.v2021_03_15_preview.models>`
            * 2021-08-15: :mod:`v2021_08_15.models<azure.mgmt.extendedlocation.v2021_08_15.models>`
+           * 2021-08-31-preview: :mod:`v2021_08_31_preview.models<azure.mgmt.extendedlocation.v2021_08_31_preview.models>`
         """
         if api_version == '2021-03-15-preview':
             from .v2021_03_15_preview import models
             return models
         elif api_version == '2021-08-15':
             from .v2021_08_15 import models
+            return models
+        elif api_version == '2021-08-31-preview':
+            from .v2021_08_31_preview import models
             return models
         raise ValueError("API version {} is not available".format(api_version))
 
@@ -108,14 +110,30 @@ class CustomLocations(MultiApiClientMixin, _SDKClient):
 
            * 2021-03-15-preview: :class:`CustomLocationsOperations<azure.mgmt.extendedlocation.v2021_03_15_preview.operations.CustomLocationsOperations>`
            * 2021-08-15: :class:`CustomLocationsOperations<azure.mgmt.extendedlocation.v2021_08_15.operations.CustomLocationsOperations>`
+           * 2021-08-31-preview: :class:`CustomLocationsOperations<azure.mgmt.extendedlocation.v2021_08_31_preview.operations.CustomLocationsOperations>`
         """
         api_version = self._get_api_version('custom_locations')
         if api_version == '2021-03-15-preview':
             from .v2021_03_15_preview.operations import CustomLocationsOperations as OperationClass
         elif api_version == '2021-08-15':
             from .v2021_08_15.operations import CustomLocationsOperations as OperationClass
+        elif api_version == '2021-08-31-preview':
+            from .v2021_08_31_preview.operations import CustomLocationsOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'custom_locations'".format(api_version))
+        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
+
+    @property
+    def resource_sync_rules(self):
+        """Instance depends on the API version:
+
+           * 2021-08-31-preview: :class:`ResourceSyncRulesOperations<azure.mgmt.extendedlocation.v2021_08_31_preview.operations.ResourceSyncRulesOperations>`
+        """
+        api_version = self._get_api_version('resource_sync_rules')
+        if api_version == '2021-08-31-preview':
+            from .v2021_08_31_preview.operations import ResourceSyncRulesOperations as OperationClass
+        else:
+            raise ValueError("API version {} does not have operation group 'resource_sync_rules'".format(api_version))
         return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
 
     def close(self):
