@@ -5,4 +5,4 @@
 # license information.
 # --------------------------------------------------------------------------
 
-VERSION = "1.0.0"
+VERSION = "1.1.0"
