@@ -11,16 +11,17 @@
 
 from typing import Any, Optional, TYPE_CHECKING
 
-from azure.core.pipeline.transport import AsyncHttpResponse, HttpRequest
+from msrest import Deserializer, Serializer
+
 from azure.mgmt.core import AsyncARMPipelineClient
 from azure.profiles import KnownProfiles, ProfileDefinition
 from azure.profiles.multiapiclient import MultiApiClientMixin
-from msrest import Deserializer, Serializer
 
 from ._configuration import CustomLocationsConfiguration
 
 if TYPE_CHECKING:
     # pylint: disable=unused-import,ungrouped-imports
+    from azure.core.credentials import TokenCredential
     from azure.core.credentials_async import AsyncTokenCredential
 
 class _SDKClient(object):
@@ -68,12 +69,10 @@ class CustomLocations(MultiApiClientMixin, _SDKClient):
         credential: "AsyncTokenCredential",
         subscription_id: str,
         api_version: Optional[str] = None,
-        base_url: Optional[str] = None,
+        base_url: str = "https://management.azure.com",
         profile: KnownProfiles = KnownProfiles.default,
         **kwargs  # type: Any
     ) -> None:
-        if not base_url:
-            base_url = 'https://management.azure.com'
         self._config = CustomLocationsConfiguration(credential, subscription_id, **kwargs)
         self._client = AsyncARMPipelineClient(base_url=base_url, config=self._config, **kwargs)
         super(CustomLocations, self).__init__(
@@ -91,12 +90,16 @@ class CustomLocations(MultiApiClientMixin, _SDKClient):
 
            * 2021-03-15-preview: :mod:`v2021_03_15_preview.models<azure.mgmt.extendedlocation.v2021_03_15_preview.models>`
            * 2021-08-15: :mod:`v2021_08_15.models<azure.mgmt.extendedlocation.v2021_08_15.models>`
+           * 2021-08-31-preview: :mod:`v2021_08_31_preview.models<azure.mgmt.extendedlocation.v2021_08_31_preview.models>`
         """
         if api_version == '2021-03-15-preview':
             from ..v2021_03_15_preview import models
             return models
         elif api_version == '2021-08-15':
             from ..v2021_08_15 import models
+            return models
+        elif api_version == '2021-08-31-preview':
+            from ..v2021_08_31_preview import models
             return models
         raise ValueError("API version {} is not available".format(api_version))
 
@@ -106,14 +109,30 @@ class CustomLocations(MultiApiClientMixin, _SDKClient):
 
            * 2021-03-15-preview: :class:`CustomLocationsOperations<azure.mgmt.extendedlocation.v2021_03_15_preview.aio.operations.CustomLocationsOperations>`
            * 2021-08-15: :class:`CustomLocationsOperations<azure.mgmt.extendedlocation.v2021_08_15.aio.operations.CustomLocationsOperations>`
+           * 2021-08-31-preview: :class:`CustomLocationsOperations<azure.mgmt.extendedlocation.v2021_08_31_preview.aio.operations.CustomLocationsOperations>`
         """
         api_version = self._get_api_version('custom_locations')
         if api_version == '2021-03-15-preview':
             from ..v2021_03_15_preview.aio.operations import CustomLocationsOperations as OperationClass
         elif api_version == '2021-08-15':
             from ..v2021_08_15.aio.operations import CustomLocationsOperations as OperationClass
+        elif api_version == '2021-08-31-preview':
+            from ..v2021_08_31_preview.aio.operations import CustomLocationsOperations as OperationClass
         else:
             raise ValueError("API version {} does not have operation group 'custom_locations'".format(api_version))
+        return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
+
+    @property
+    def resource_sync_rules(self):
+        """Instance depends on the API version:
+
+           * 2021-08-31-preview: :class:`ResourceSyncRulesOperations<azure.mgmt.extendedlocation.v2021_08_31_preview.aio.operations.ResourceSyncRulesOperations>`
+        """
+        api_version = self._get_api_version('resource_sync_rules')
+        if api_version == '2021-08-31-preview':
+            from ..v2021_08_31_preview.aio.operations import ResourceSyncRulesOperations as OperationClass
+        else:
+            raise ValueError("API version {} does not have operation group 'resource_sync_rules'".format(api_version))
         return OperationClass(self._client, self._config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
 
     async def close(self):
