@@ -43,7 +43,7 @@ async def run_sample():
     # Let's create a secret holding storage account credentials.
     # if the secret already exists in the Key Vault, then a new version of the secret is created.
     print("\n.. Create Secret")
-    secret = await client.set_secret("backupRestoreSecretName", "backupRestoreSecretValue")
+    secret = await client.set_secret("backupRestoreSecretNameAsync", "backupRestoreSecretValue")
     print("Secret with name '{0}' created with value '{1}'".format(secret.name, secret.value))
 
     # Backups are good to have, if in case secrets gets deleted accidentally.
