@@ -4,4 +4,4 @@
 # license information.
 # --------------------------------------------------------------------------
 
-VERSION = "12.9.0"
+VERSION = "12.10.0b1"
