@@ -9,4 +9,4 @@
 # regenerated.
 # --------------------------------------------------------------------------
 
-VERSION = "12.4.0"
+VERSION = "12.5.0b1"
