@@ -12,13 +12,12 @@ from typing import (
     cast,
     TYPE_CHECKING
 )
-from functools import partial
 from azure.core.paging import ItemPaged
 from azure.core.tracing.decorator import distributed_trace
 from azure.core.exceptions import HttpResponseError
 from azure.core.credentials import AzureKeyCredential
 from ._base_client import TextAnalyticsClientBase
-from ._lro import AnalyzeActionsLROPoller, AnalyzeHealthcareEntitiesLROPoller
+from ._lro import AnalyzeActionsLROPoller, AnalyzeHealthcareEntitiesLROPoller, TextAnalyticsLROPoller
 from ._request_handlers import (
     _validate_input,
     _determine_action_type,
@@ -35,6 +34,7 @@ from ._response_handlers import (
     pii_entities_result,
     healthcare_paged_result,
     analyze_paged_result,
+    _get_result_from_continuation_token
 )
 
 from ._lro import (
@@ -596,8 +596,9 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
         except HttpResponseError as error:
             return process_http_response_error(error)
 
+    # pylint: disable=unused-argument
     def _healthcare_result_callback(
-        self, doc_id_order, raw_response, deserialized, headers, show_stats=False
+        self, raw_response, deserialized, doc_id_order, task_id_order=None, show_stats=False, bespoke=False
     ):
         if deserialized is None:
             models = self._client.models(api_version=self._api_version)
@@ -609,7 +610,6 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
             self._client.analyze_text_job_status if is_language_api(self._api_version) else self._client.health_status,
             raw_response,
             deserialized,
-            headers,
             show_stats=show_stats,
         )
 
@@ -700,26 +700,20 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
         display_name = kwargs.pop("display_name", None)
 
         if continuation_token:
-            def get_result_from_cont_token(initial_response, pipeline_response):
-                doc_id_order = initial_response.context.options["doc_id_order"]
-                show_stats = initial_response.context.options["show_stats"]
-                return self._healthcare_result_callback(
-                    doc_id_order, pipeline_response, None, {}, show_stats=show_stats
-                )
-
             return cast(
                 AnalyzeHealthcareEntitiesLROPoller[
                     ItemPaged[Union[AnalyzeHealthcareEntitiesResult, DocumentError]]
                 ],
-                AnalyzeHealthcareEntitiesLROPoller.from_continuation_token(
-                    polling_method=AnalyzeHealthcareEntitiesLROPollingMethod(
+                _get_result_from_continuation_token(
+                    self._client._client,  # pylint: disable=protected-access
+                    continuation_token,
+                    AnalyzeHealthcareEntitiesLROPoller,
+                    AnalyzeHealthcareEntitiesLROPollingMethod(
                         text_analytics_client=self._client,
                         timeout=polling_interval,
                         **kwargs
                     ),
-                    client=self._client._client,  # pylint: disable=protected-access
-                    deserialization_callback=get_result_from_cont_token,
-                    continuation_token=continuation_token
+                    self._healthcare_result_callback
                 )
             )
 
@@ -727,8 +721,8 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
         doc_id_order = [doc.get("id") for doc in docs]
         my_cls = kwargs.pop(
             "cls",
-            partial(
-                self._healthcare_result_callback, doc_id_order, show_stats=show_stats
+            lambda pipeline_response, deserialized, _: self._healthcare_result_callback(
+                pipeline_response, deserialized, doc_id_order, show_stats=show_stats
             ),
         )
         models = self._client.models(api_version=self._api_version)
@@ -1030,7 +1024,7 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
             return process_http_response_error(error)
 
     def _analyze_result_callback(
-        self, doc_id_order, task_order, raw_response, deserialized, headers, show_stats=False
+        self, raw_response, deserialized, doc_id_order, task_id_order=None, show_stats=False, bespoke=False
     ):
 
         if deserialized is None:
@@ -1039,12 +1033,12 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
             deserialized = response_cls.deserialize(raw_response)
         return analyze_paged_result(
             doc_id_order,
-            task_order,
+            task_id_order,
             self._client.analyze_text_job_status if is_language_api(self._api_version) else self._client.analyze_status,
             raw_response,
             deserialized,
-            headers,
             show_stats=show_stats,
+            bespoke=bespoke
         )
 
     @distributed_trace
@@ -1166,27 +1160,23 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
         show_stats = kwargs.pop("show_stats", None)
         polling_interval = kwargs.pop("polling_interval", 5)
         language = language_arg if language_arg is not None else self._default_language
+        poller_cls = kwargs.pop("poller_cls", AnalyzeActionsLROPoller)
+        bespoke = kwargs.pop("bespoke", False)
 
         if continuation_token:
-            def get_result_from_cont_token(initial_response, pipeline_response):
-                doc_id_order = initial_response.context.options["doc_id_order"]
-                task_id_order = initial_response.context.options["task_id_order"]
-                show_stats = initial_response.context.options["show_stats"]
-                return self._analyze_result_callback(
-                    doc_id_order, task_id_order, pipeline_response, None, {}, show_stats=show_stats
-                )
-
             return cast(
                 AnalyzeActionsResponse,
-                AnalyzeActionsLROPoller.from_continuation_token(
-                    polling_method=AnalyzeActionsLROPollingMethod(
+                _get_result_from_continuation_token(
+                    self._client._client,  # pylint: disable=protected-access
+                    continuation_token,
+                    AnalyzeActionsLROPoller,
+                    AnalyzeActionsLROPollingMethod(
                         text_analytics_client=self._client,
                         timeout=polling_interval,
                         **kwargs
                     ),
-                    client=self._client._client,  # pylint: disable=protected-access
-                    deserialization_callback=get_result_from_cont_token,
-                    continuation_token=continuation_token
+                    self._analyze_result_callback,
+                    bespoke
                 )
             )
 
@@ -1206,6 +1196,18 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
         except AttributeError as e:
             raise TypeError("Unsupported action type in list.") from e
         task_order = [(_determine_action_type(a), a.task_name) for a in generated_tasks]
+        response_cls = kwargs.pop(
+            "cls",
+            lambda pipeline_response, deserialized, _:
+                self._analyze_result_callback(
+                    pipeline_response,
+                    deserialized,
+                    doc_id_order,
+                    task_id_order=task_order,
+                    show_stats=show_stats,
+                    bespoke=bespoke
+                ),
+        )
 
         try:
             if is_language_api(self._api_version):
@@ -1217,15 +1219,7 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
                             display_name=display_name,
                             tasks=generated_tasks
                         ),
-                        cls=kwargs.pop(
-                            "cls",
-                            partial(
-                                self._analyze_result_callback,
-                                doc_id_order,
-                                task_order,
-                                show_stats=show_stats,
-                            ),
-                        ),
+                        cls=response_cls,
                         polling=AnalyzeActionsLROPollingMethod(
                             text_analytics_client=self._client,
                             timeout=polling_interval,
@@ -1240,6 +1234,7 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
                             **kwargs
                         ),
                         continuation_token=continuation_token,
+                        poller_cls=poller_cls,
                         **kwargs
                     )
                 )
@@ -1274,15 +1269,7 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
                 AnalyzeActionsResponse,
                 self._client.begin_analyze(
                     body=analyze_body,
-                    cls=kwargs.pop(
-                        "cls",
-                        partial(
-                            self._analyze_result_callback,
-                            doc_id_order,
-                            task_order,
-                            show_stats=show_stats,
-                        ),
-                    ),
+                    cls=response_cls,
                     polling=AnalyzeActionsLROPollingMethod(
                         text_analytics_client=self._client,
                         timeout=polling_interval,
@@ -1300,5 +1287,124 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
                     **kwargs
                 )
             )
+        except HttpResponseError as error:
+            return process_http_response_error(error)
+
+    @distributed_trace
+    @validate_multiapi_args(
+        version_method_added="2022-05-01"
+    )
+    def begin_recognize_custom_entities(
+        self,
+        documents: Union[List[str], List[TextDocumentInput], List[Dict[str, str]]],
+        project_name,
+        deployment_name,
+        **kwargs: Any,
+    ) -> TextAnalyticsLROPoller[ItemPaged[Union[RecognizeCustomEntitiesResult, DocumentError]]]:
+        """Start a long-running custom named entity recognition operation.
+
+        For information on regional support of custom features and how to train a model to
+        recognize custom entities, see https://aka.ms/azsdk/textanalytics/customentityrecognition
+
+        :param documents: The set of documents to process as part of this batch.
+            If you wish to specify the ID and language on a per-item basis you must
+            use as input a list[:class:`~azure.ai.textanalytics.TextDocumentInput`] or a list of
+            dict representations of :class:`~azure.ai.textanalytics.TextDocumentInput`, like
+            `{"id": "1", "language": "en", "text": "hello world"}`.
+        :type documents:
+            list[str] or list[~azure.ai.textanalytics.TextDocumentInput] or list[dict[str, str]]
+        :param str project_name: Required. This field indicates the project name for the model.
+        :param str deployment_name: This field indicates the deployment name for the model.
+        :keyword str language: The 2 letter ISO 639-1 representation of language for the
+            entire batch. For example, use "en" for English; "es" for Spanish etc.
+            If not set, uses "en" for English as default. Per-document language will
+            take precedence over whole batch language. See https://aka.ms/talangs for
+            supported languages in Language API.
+        :keyword bool show_stats: If set to true, response will contain document level statistics.
+        :keyword bool disable_service_logs: If set to true, you opt-out of having your text input
+            logged on the service side for troubleshooting. By default, the Language service logs your
+            input text for 48 hours, solely to allow for troubleshooting issues in providing you with
+            the service's natural language processing functions. Setting this parameter to true,
+            disables input logging and may limit our ability to remediate issues that occur. Please see
+            Cognitive Services Compliance and Privacy notes at https://aka.ms/cs-compliance for
+            additional details, and Microsoft Responsible AI principles at
+            https://www.microsoft.com/ai/responsible-ai.
+        :keyword str string_index_type: Specifies the method used to interpret string offsets.
+            `UnicodeCodePoint`, the Python encoding, is the default. To override the Python default,
+            you can also pass in `Utf16CodeUnit` or `TextElement_v8`. For additional information
+            see https://aka.ms/text-analytics-offsets
+        :keyword int polling_interval: Waiting time between two polls for LRO operations
+            if no Retry-After header is present. Defaults to 5 seconds.
+        :keyword str continuation_token:
+            Call `continuation_token()` on the poller object to save the long-running operation (LRO)
+            state into an opaque token. Pass the value as the `continuation_token` keyword argument
+            to restart the LRO from a saved state.
+        :keyword str display_name: An optional display name to set for the requested analysis.
+        :return: An instance of an TextAnalyticsLROPoller. Call `result()` on the this
+            object to return a heterogeneous pageable of
+            :class:`~azure.ai.textanalytics.RecognizeCustomEntitiesResult` and
+            :class:`~azure.ai.textanalytics.DocumentError`.
+        :rtype:
+            ~azure.ai.textanalytics.TextAnalyticsLROPoller[~azure.core.paging.ItemPaged[
+            ~azure.ai.textanalytics.RecognizeCustomEntitiesResult or ~azure.ai.textanalytics.DocumentError]]
+        :raises ~azure.core.exceptions.HttpResponseError:
+
+        .. versionadded:: 2022-05-01
+            The *begin_recognize_custom_entities* client method.
+
+        .. admonition:: Example:
+
+            .. literalinclude:: ../samples/sample_recognize_custom_entities.py
+                :start-after: [START recognize_custom_entities]
+                :end-before: [END recognize_custom_entities]
+                :language: python
+                :dedent: 4
+                :caption: Recognize custom entities in a batch of documents.
+        """
+
+        continuation_token = kwargs.pop("continuation_token", None)
+        string_index_type = kwargs.pop("string_index_type", self._string_index_type_default)
+        disable_service_logs = kwargs.pop("disable_service_logs", None)
+        polling_interval = kwargs.pop("polling_interval", 5)
+
+        if continuation_token:
+            return cast(
+                TextAnalyticsLROPoller[ItemPaged[Union[RecognizeCustomEntitiesResult, DocumentError]]],
+                _get_result_from_continuation_token(
+                    self._client._client,  # pylint: disable=protected-access
+                    continuation_token,
+                    TextAnalyticsLROPoller,
+                    AnalyzeActionsLROPollingMethod(
+                        text_analytics_client=self._client,
+                        timeout=polling_interval,
+                        **kwargs
+                    ),
+                    self._analyze_result_callback,
+                    bespoke=True
+                )
+            )
+
+        try:
+            return cast(
+                TextAnalyticsLROPoller[
+                    ItemPaged[Union[RecognizeCustomEntitiesResult, DocumentError]]
+                ],
+                self.begin_analyze_actions(
+                    documents,
+                    actions=[
+                        RecognizeCustomEntitiesAction(
+                            project_name=project_name,
+                            deployment_name=deployment_name,
+                            string_index_type=string_index_type,
+                            disable_service_logs=disable_service_logs
+                        )
+                    ],
+                    polling_interval=polling_interval,
+                    poller_cls=TextAnalyticsLROPoller,
+                    bespoke=True,
+                    **kwargs
+                )
+            )
+
         except HttpResponseError as error:
             return process_http_response_error(error)
