@@ -9,7 +9,7 @@ FILE: sample_recognize_custom_entities.py
 
 DESCRIPTION:
     This sample demonstrates how to recognize custom entities in documents.
-    Recognizing custom entities is available as an action type through the begin_analyze_actions API.
+    Recognizing custom entities is also available as an action type through the begin_analyze_actions API.
 
     For information on regional support of custom features and how to train a model to
     recognize custom entities, see https://aka.ms/azsdk/textanalytics/customentityrecognition
@@ -29,11 +29,9 @@ import os
 
 
 def sample_recognize_custom_entities():
+    # [START recognize_custom_entities]
     from azure.core.credentials import AzureKeyCredential
-    from azure.ai.textanalytics import (
-        TextAnalyticsClient,
-        RecognizeCustomEntitiesAction,
-    )
+    from azure.ai.textanalytics import TextAnalyticsClient
 
     endpoint = os.environ["AZURE_LANGUAGE_ENDPOINT"]
     key = os.environ["AZURE_LANGUAGE_KEY"]
@@ -55,18 +53,14 @@ def sample_recognize_custom_entities():
     with open(path_to_sample_document) as fd:
         document = [fd.read()]
 
-    poller = text_analytics_client.begin_analyze_actions(
+    poller = text_analytics_client.begin_recognize_custom_entities(
         document,
-        actions=[
-            RecognizeCustomEntitiesAction(
-                project_name=project_name, deployment_name=deployment_name
-            ),
-        ],
+        project_name=project_name,
+        deployment_name=deployment_name
     )
 
     document_results = poller.result()
-    for result in document_results:
-        custom_entities_result = result[0]  # first document, first result
+    for custom_entities_result in document_results:
         if not custom_entities_result.is_error:
             for entity in custom_entities_result.entities:
                 print(
@@ -80,6 +74,7 @@ def sample_recognize_custom_entities():
                     custom_entities_result.code, custom_entities_result.message
                 )
             )
+    # [END recognize_custom_entities]
 
 
 if __name__ == "__main__":
